requirements.txt
streamlit
easyocr
Pillow
opencv-python-headless  # opencv-python 대신 opencv-python-headless 사용
numpy
seaborn
openpyxl
python-docx
python-pptx
olefile